fix(HTD-11): move input to modules
