+++ src/app/page.tsx
@@ -3,7 +3,6 @@ import { Button } from '@/components/elements/button';
 import { Container } from '@/components/elements/container';
 import { Heading } from '@/components/elements/heading';
 import { Sparkle, Alien, Logo, CaretDown, RocketLaunch } from '@/components/elements/icon';
-import { Input } from '@/components/elements/input';
 import ShowCase from '@/components/elements/showcase';
 import { SquareIcon } from '@/components/elements/square-icon';
 import { Text } from '@/components/elements/text';
@@ -11,6 +10,7 @@ import { ThemeToggle } from '@/components/elements/theme-toggle';
 import { Tooltip } from '@/components/elements/tooltip';
 import { Dropdown } from '@/components/modules/dropdown';
 import type { DropdownOptionProps } from '@/components/modules/dropdown/dropdown.types';
+import { Input } from '@/components/modules/input';
 
 export default async function Home() {
   const dropdownItems: DropdownOptionProps[] = [
